refactor(tests): rename namespace SharedStepDefinitions to Common
