+++ RemoteControlledProcess.Acceptance.Tests/Steps/Common/ValidateCoverageStepDefinitions.cs
@@ -7,7 +7,7 @@ using TechTalk.SpecFlow;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace RemoteControlledProcess.Acceptance.Tests.Steps.SharedStepDefinitions;
+namespace RemoteControlledProcess.Acceptance.Tests.Steps.Common;
 
 [Binding]
 public partial class ValidateCoverageStepDefinitions
