+++ RemoteControlledProcess.Acceptance.Tests/Steps/PassEnvironmentVariablesStepDefinitions.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using RemoteControlledProcess.Acceptance.Tests.Steps.SharedStepDefinitions;
+using RemoteControlledProcess.Acceptance.Tests.Steps.Common;
 using TechTalk.SpecFlow;
 using Xunit;
